Advanced Results Middleware
Advanced Results(filter, pagination, etc.) Middleware for bootcamp and courses
